tried to do homework stuck on distance
As with the last mapping assignment I cannot get the variable to the
.distanceTo and thus cannot complete antything else. I believe this is
the result of knot knowing how to use these functions and while I
understand the api is telling me what it can do, how to execute those
functions (what format data needs to be in,etc. seems to be lost on me).
Specifally, I am wondering if I need to extract the latlng in a specific
way--however, I have not idea why passing the coordinates
function.geometry.coordinates wouldn't be the same. but apparently its
not a function